Add check_for_audit_tables method
