+++ spec/audit_table_spec.rb
@@ -27,9 +27,14 @@ describe AuditTables do
   end
 
   describe 'Audit Tables' do
+
+    after(:all) do
+      ActiveRecord::Base.connection.close
+    end
+
     context '.create_audit_table_for' do
       before { setup_database }
-
+      
       it 'database is audit_tables_test' do
         expect(ActiveRecord::Base.connection_config[:database]).to match(/audit_tables_test/)
       end
