+++ spec/spec_helper.rb
@@ -2,15 +2,8 @@
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 
-require 'rubygems'
-require 'bundler'
-
-Bundler.setup(:default, :development)
-
 require 'active_record'
-require 'active_support'
 require 'audit_tables'
-require 'rspec'
 require 'faker'
 
 ActiveRecord::Base.establish_connection(adapter: 'postgresql', database: 'postgres')
@@ -20,6 +13,3 @@ ActiveRecord::Base.connection.execute('CREATE DATABASE audit_tables_test;')
 ActiveRecord::Base.connection.close
 
 ActiveRecord::Base.establish_connection(adapter: 'postgresql', database: 'audit_tables_test')
-
-RSpec.configure do |config|
-end
